backend/requirements.txt
# AeroHire Backend Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25

# Data Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# File Upload
python-multipart==0.0.9

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Docker SDK (for sandbox code execution)
docker==7.0.0

# PDF Parsing (for resume extraction)
pdfplumber==0.10.4

# Google AI (for Glass Box reasoning)
google-generativeai==0.4.0
